Work on ADD

+++ Moira/MoiraUtils.cpp
@@ -9,21 +9,23 @@
 
 #include "MoiraUtils.h"
 
-template<> u32 CLIP<Byte>(u32 data) { return data & 0xff; }
-template<> u32 CLIP<Word>(u32 data) { return data & 0xffff; }
-template<> u32 CLIP<Long>(u32 data) { return data & 0xffffffff; }
+template<> u32 CLIP<Byte>(u64 data) { return data & 0xff; }
+template<> u32 CLIP<Word>(u64 data) { return data & 0xffff; }
+template<> u32 CLIP<Long>(u64 data) { return data & 0xffffffff; }
 
-template<> u32 MASK<Byte>() { return 0xff; }
-template<> u32 MASK<Word>() { return 0xffff; }
-template<> u32 MASK<Long>() { return 0xffffffff; }
+/*
+template<> u64 MASK<Byte>() { return 0xff; }
+template<> u64 MASK<Word>() { return 0xffff; }
+template<> u64 MASK<Long>() { return 0xffffffff; }
+*/
 
-template<> bool MSBIT<Byte>(u32 data) { return data & 0x80; }
-template<> bool MSBIT<Word>(u32 data) { return data & 0x8000; }
-template<> bool MSBIT<Long>(u32 data) { return data & 0x80000000; }
+template<> bool MSBIT<Byte>(u64 data) { return data & 0x80; }
+template<> bool MSBIT<Word>(u64 data) { return data & 0x8000; }
+template<> bool MSBIT<Long>(u64 data) { return data & 0x80000000; }
 
-template<> u32 SIGN<Byte>(u32 data) { return  (i8)data; }
-template<> u32 SIGN<Word>(u32 data) { return (i16)data; }
-template<> u32 SIGN<Long>(u32 data) { return (i32)data; }
+template<> u32 SIGN<Byte>(u64 data) { return  (i8)data; }
+template<> u32 SIGN<Word>(u64 data) { return (i16)data; }
+template<> u32 SIGN<Long>(u64 data) { return (i32)data; }
 
 template<> u8 BYTES<Byte>() { return 1; }
 template<> u8 BYTES<Word>() { return 2; }
